perf(k3): block-parallel router top-k replaces the serial TileLang scan
The TileLang router_topk_batched kernel ran the whole biased top-k as a
serial scan on thread 0 — TOPK x E iterations per row, 65.7us per launch
at 896 experts, ~6ms of every EP16 verify round across the 92 expert
layers. Replace it with a hand-written CUDA kernel whose selection is a
block-parallel argmax per round while every arithmetic step keeps the
retired kernel's spelling, so the outputs are bit-identical: f32 sigmoid
via plain expf, bias added in f32, strict less-than comparisons with the
lowest-index tie-break (each thread scans its stride ascending; the
reduction prefers the strictly greater value and breaks equal values to
the lower index, which is exactly the serial first-match), the
denominator accumulated in selection order, and the division-then-scale
weight normalization.

The TileLang factory, its generator plan and the AOT launcher entry are
retired; batch and expert count become plain runtime launch values (no
per-bucket instantiation).

Gates on the pruned 224-expert checkpoint: golden decode 13/13 (mega +
masked chain — any selection divergence would blow past the 2.0-ULP
noise floor), spec_verify 6/6, lib tests green.

Perf: 65.7us -> 10.3us per launch (bucket-one golden replay capture,
240 launches, max 11.2us); an EP16 verify step's 92 router launches drop
from ~6.0ms to ~0.95ms.
