+++ pegainfer-kernels/src/ops/k3/mod.rs
@@ -6,6 +6,7 @@ mod flash_mla_prefill;
 mod mega_moe;
 mod mla_paged;
 mod moe_chain;
+mod router_topk;
 
 pub use deepgemm::*;
 pub use flash_kda::*;
@@ -13,3 +14,4 @@ pub use flash_mla_prefill::*;
 pub use mega_moe::*;
 pub use mla_paged::*;
 pub use moe_chain::*;
+pub use router_topk::*;
